adjusted initial & solution
